feat: implemented coil selection in measurement settings #trench-105
