+++ frontend/src/app/configuration/measurement-settings/components/management/measurement-settings.component.spec.ts
